CARDS-1886: Replace Torch import with Clarity import Move the Torch Import code into a separate module
