+++ modules/torch-import/src/main/java/io/uhndata/cards/torch/internal/ImportTask.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package io.uhndata.cards.proms.internal.importer;
+package io.uhndata.cards.torch.internal;
 
 import java.io.BufferedReader;
 import java.io.IOException;
