+++ modules/torch-import/src/main/java/io/uhndata/cards/torch/internal/PatientLocalStorage.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package io.uhndata.cards.proms.internal.importer;
+package io.uhndata.cards.torch.internal;
 
 import java.io.IOException;
 import java.text.ParseException;
